requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-binance==1.0.19
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.4
pyOpenSSL==23.2.0